005_x_data_pipeline_using_airflow/requirements.txt
pandas
tweepy
s3fs
python-dotenv